+++ include/application.hpp
@@ -87,6 +87,8 @@ namespace VIPR_Emulator
 			friend void machine_options_ram_in_kb_input_complete(GUI::Value &obj, void *userdata);
 			friend void machine_options_rom_file_input_complete(GUI::Input &obj, void *userdata);
 
+			friend void expansion_board_options_activate(GUI::Menu &obj, void *userdata);
+
 			friend void machine_memory_transfer_left(GUI::Menu &obj, void *userdata);
 			friend void machine_memory_transfer_right(GUI::Menu &obj, void *userdata);
 			friend void machine_memory_transfer_activate(GUI::Menu &obj, void *userdata);
@@ -108,7 +110,7 @@ namespace VIPR_Emulator
 			std::map<HexKey, SDL_Scancode> Hex_KeyMap;
 			std::multimap<char, ScancodeModData> Printable_KeyMap;
 			COSMAC_VIP System;
-			GUI::Menu MainMenu, MachineOptionsMenu, MachineMemoryTransferMenu, EmulatorOptionsMenu;
+			GUI::Menu MainMenu, MachineOptionsMenu, ExpansionBoardOptionsMenu, MachineMemoryTransferMenu, EmulatorOptionsMenu;
 			GUI::Menu *CurrentMenu;
 			GUI::ElementData *InputFocus;
 			bool exit;
@@ -152,6 +154,12 @@ namespace VIPR_Emulator
 	void machine_options_activate(GUI::Menu &obj, void *userdata);
 	void machine_options_ram_in_kb_input_complete(GUI::Value &obj, void *userdata);
 	void machine_options_rom_file_input_complete(GUI::Input &obj, void *userdata);
+
+	void expansion_board_options_up(GUI::Menu &obj, void *userdata);
+	void expansion_board_options_down(GUI::Menu &obj, void *userdata);
+	void expansion_board_options_left(GUI::Menu &obj, void *userdata);
+	void expansion_board_options_right(GUI::Menu &obj, void *userdata);
+	void expansion_board_options_activate(GUI::Menu &obj, void *userdata);
 
 	void machine_memory_transfer_up(GUI::Menu &obj, void *userdata);
 	void machine_memory_transfer_down(GUI::Menu &obj, void *userdata);