+++ include/audio.hpp
@@ -1,0 +1,12 @@
+#ifndef _AUDIO_HPP_
+#define _AUDIO_HPP_
+
+#include <cstdint>
+#include <array>
+
+namespace VIPR_Emulator
+{
+	using AudioFrame = std::array<int, 4096>;
+}
+
+#endif